Move vignetteHelper to utils
