+++ source/utils/vignetteHelper.js
